Feature/up 1893 upgrade and cassandra (#3)
* Updated to the latest express kafka
Added an ExecutionContextHelper to have a fixed amount of threads amongst the different parts of the services
removed an unused prometheus metric initialized at the boot
Updated the tests to be able to stop the kafka consumer / producer once the tests are over (BUG)
Changed the way the redis shutdown is made. For this

* Fixed the tests

* Reduced log level in test to make it more readeable

* Integrated cassandra

* Cleanup
Fixed tests: they are now passing with mvn test

* Removed non needed class

* Removed possibility to disable filtering if env = prod

* added error logigng when cassandra refuses to start

* added error logigng when cassandra refuses to start

* No idea why CI is failing to start cassandra, maybe this will make it clearer?

* Small code cleanup

* Small code cleanup

* Fixed application.conf files
FilterError deserializer now deserialize strings
Changed where different messages are sent. The topic can now be configured through env variables and are no longer fixed
Now parsing and sending strings instead of byte arrays

* Added http headers to replay attack

* Added maven scaliform plugin, hence the huge amount of code formatting done
Refactor -> removed Messages class and merged it with the Values class
    -> fake filter service are now defined in their own class and not randomly at the end of the test suits
    -> A redis mock is now put in the proper place
    -> Renamed an exception

* Removed non needed classes

* Removed use of the kill command that made the CI unhappy
Made tests slightly faster

Co-authored-by: sharkgua <benoit.george@ubirch.com>